feat: handle multipart form-data body requests

+++ src/writer/model/pinnedFile.go
@@ -4,7 +4,7 @@ import "time"
 
 type PinnedFile struct {
 	IpfsHash     string
-	PinSize      int8
+	PinSize      int64
 	Timestamp    time.Time
 	IsDuplicated *bool
 }
